Refactoring monitoring functionality
Add test for monitoring devices. Reorganise code from the "main" package
into the "device" package.
Use the io.Writer as an interface to be able to test printed messages.
Create an abstraction called Scanner interface to be able to test device
related events without interacting with real hardware using Udev.
Create an abstraction to represent an event happened to a device called
SecurityTokenEvent.