Implement authentication using PAM
Create two interfaces for providing authentication: Authenticator and
CurretUserAuthenticator
First needs a username, the second will use the actual user.
PamAuthenticator implements the first, while OsCurrentUserAuthenticator
the second.
In addition to that a UserProvider interface has been added for
obtaining the actual user in an abstract way. CurrentUserProvider
implements it by taking the current user from the Operating System.

All of the above is wired together in an action called
AuthenticateCurrentUserAction,
which is simply a callback function.